+++ src/data-access/repositories/__mocks__/aws-sdk/clients/dynamodb.ts
